requirements_installation.txt
prefect==2.0b3
elasticsearch==8.1.2
mysql-connector-python==8.0.28
